Tested hoster to work with new libraries

+++ src/main/java/com/github/vmorev/crawler/tools/Cleaner.java
@@ -5,14 +5,18 @@ import com.github.vmorev.amazon.S3Bucket;
 import com.github.vmorev.amazon.SDBDomain;
 import com.github.vmorev.amazon.SQSQueue;
 
+import java.util.Arrays;
+import java.util.List;
+
 /**
  * User: valentin
  * Date: 10.02.13
  */
 public class Cleaner {
+    private static final List<String> s3exceptions = Arrays.asList("crawler-articles");
 
     public static void main(String[] args) throws Exception {
-        //cleanS3();
+        cleanS3();
         cleanSQS();
         cleanSDB();
     }
@@ -20,8 +24,10 @@ public class Cleaner {
     private static void cleanS3() {
         for (Bucket bucket : S3Bucket.listBuckets()) {
             try {
-                new S3Bucket(bucket.getName()).deleteBucket();
-                System.out.println("SUCCESS. S3 bucket was deleted " + bucket.getName());
+                if (!s3exceptions.contains(bucket.getName())) {
+                    new S3Bucket(bucket.getName()).deleteBucket();
+                    System.out.println("SUCCESS. S3 bucket was deleted " + bucket.getName());
+                }
             } catch (Exception e) {
                 System.out.println("FAIL. S3 bucket was NOT deleted " + bucket.getName());
             }
